+++ lib/flows/shared_context_pipeline/dsl/tracks.rb
@@ -16,15 +16,15 @@ module Flows
 
         attr_reader :tracks
 
-        def step(name, router_def = DEFAULT_ROUTER_DEF, &lambda)
+        def step(name, router_def = DEFAULT_ROUTER_DEF, body: nil)
           tracks.add_step(
-            Step.new(name: name, lambda: lambda, router_def: router_def)
+            Step.new(name: name, body: body, router_def: router_def)
           )
         end
 
-        def mut_step(name, router_def = DEFAULT_ROUTER_DEF, &lambda)
+        def mut_step(name, router_def = DEFAULT_ROUTER_DEF, body: nil)
           tracks.add_step(
-            MutationStep.new(name: name, lambda: lambda, router_def: router_def)
+            MutationStep.new(name: name, body: body, router_def: router_def)
           )
         end