+++ lib/flows/shared_context_pipeline/step.rb
@@ -3,12 +3,22 @@ module Flows
     EMPTY_ARRAY = [].freeze
 
     # @api private
-    Step = Struct.new(:name, :lambda, :router_def, :next_step, keyword_init: true) do
+    Step = Struct.new(:name, :body, :router_def, :next_step, keyword_init: true) do
+      # :reek:ManualDispatch
+      def initialize(name:, body: nil, **rest)
+        if name.respond_to?(:call)
+          body = name
+          name = "#{body}+Step-Object-ID-#{object_id}"
+        end
+
+        super(name: name, body: body, **rest)
+      end
+
       def to_node(pipeline_class)
         klass = self.class
 
         Flows::Flow::Node.new(
-          body: lambda || pipeline_class.method(name),
+          body: body || pipeline_class.method(name),
           router: router_def.to_router(next_step),
           meta: { name: name },
           preprocessor: klass::NODE_PREPROCESSOR,